Move DiscordEvent into core, its not an actual event
