+++ app/Discord/Fun/Events/BanKickCounter.php
@@ -2,7 +2,7 @@
 
 namespace App\Discord\Fun\Events;
 
-use App\Discord\Core\Events\DiscordEvent;
+use App\Discord\Core\DiscordEvent;
 use App\Discord\Core\Models\DiscordUser;
 use App\Discord\Fun\Models\BanCounter;
 use App\Discord\Fun\Models\KickCounter;
