+++ app/Discord/Logger/Events/GuildBanRemove.php
@@ -2,12 +2,11 @@
 
 namespace App\Discord\Logger\Events;
 
-use App\Discord\Core\Events\DiscordEvent;
+use App\Discord\Core\DiscordEvent;
 use App\Discord\Logger\Enums\LogSetting;
 use Discord\Discord;
 use Discord\Parts\Guild\Ban;
 use Discord\WebSockets\Event;
-use Exception;
 
 class GuildBanRemove extends DiscordEvent
 {
